fix(opengraph): Override open graph urls for RECAP pdfs
These are a special case where we redirect opengraph bots to /recap/og-lookup/?file_path=/recap/the/path so that they can get their opengraph content.
Previously, this page was using the og:url from the base.html template, which doesn't include the query parameters.
This means that clients who depend on the og:url property for the click action on an opengraph card (e.g. Mastodon) were getting URLs that 404ed when you click the preview card but not the link in
the content that generated it.
The og:url would always be `courtlistener.com/recap/og-lookup/` which of course 404s since the query param it's expecting isn't present.

To fix, override the template on recap_document.html.
Pass an `og_file_path` argument when we do the redirect and check for it in the template.