fix(build): link to poetry in /usr/local/bin/
After #3110, which manually installed Poetry using `pip`,
the binary was no longer in `$PATH` and needed to be
addressed directly. This is fine so far as the Dockerfile
is concerned, but it is inconvenient when an implementer
wants to interact with `poetry` using `docker exec`.

Resolve the issue by creating a `poetry` symlink accessible
via `$PATH`.