chore(build): drop use of github archive for `build-prod`
When the archive download approach was introduced
(in #2002), both image building and deployment
were done locally and there was a risk that
an unclean working copy could cause WIP code to
be deployed.

Today, responsibility for building (#2118) and
deployment (#2169) have moved to GitHub Actions.
This is the best practice and has made the
archive download safeguard unnecessary.

Change the `build-prod` target to use the Docker
build context rather than a downloaded copy of
the `courtlistener` repository. This is more
conventional, respects the `.dockerignore` file,
and removes a difference between the `build-dev`
and `build-prod` targets

If it is necessary to build or push from a local
machine, downloading the archive and building the
image within that (clean) context will achieve
the same thing:
```
% cd /tmp
% curl -Lo courtlistener.tar.gz https://github.com/freelawproject/courtlistener/archive/${GIT_SHA}.tar.gz
% tar xvfz courtlistener.tar.gz && \
  mv courtlistener-${GIT_SHA} courtlistener && \
  rm -r courtlistener.tar.gz
% cd courtlistener
% make push --file docker/django/Makefile VERSION=${GIT_SHA}
```