recap: Relax `pacer_case_id` dash/hyphen constraint
Allow dashes in `pacer_case_id` as long as there is not a single one.
The concern here is the a single dash (aka hyphen) can look like a
docket number, and there was a point in time where this field was
overloaded with appellate docket numbers.

Explain a little better why this check exists.

Add a test.