fix(lint): Fix type for retry decorator
Added a tuple of type exceptions to retry decorator

It appears that we do not run mypy on cl.scrapers
either by design or by accident but the decorator
specified a single type and not many.  We mainly use
a tuple of exceptions with the retry so I added
it as the correct type.

We should probably update the mypy checker