chore(build): housekeeping for `cl-postgresql` and `cl-es` TLS
Strict verification of certificates in development isn't something
that is necessary, but it is still productive. Matching or
exceeding the rigor required by production can shake out
configuration issues before they reach production.

With a few adjustments, we can provide `cl-django` the information
it needs to verify the `cl-postgresql` and `cl-es` services.

- Remove unsafe default for `ELASTICSEARCH_CA_CERT` and instead
  specify this value in `docker-compose.yml`.
- Create `ELASTICSEARCH_VERIFY_CERT` and use it to configure
  certificate verification in development.
- Provide `cl-postgres.crt` and `cl-es.crt` to `cl-django` and
  `cl-celery`