chore(docker): correct `CL_BASE_DIR` interpolation
The entire source that is specified for the volume at
`/opt/courtlistener` is interpolated:

```
${CL_BASE_DIR:-../../../courtlistener}:/opt/courtlistener
```

This appears to have brought about a misapprehension of what
should be interpolated: in 1a57939f a files full path was the
default value for `CL_BASE_DIR`:

```
- ${CL_BASE_DIR:-../../../courtlistener/cl/search/elasticsearch_files/synonyms_en.txt}:/usr/share/elasticsearch/config/synonyms_en.txt
- ${CL_BASE_DIR:-../../../courtlistener/cl/search/elasticsearch_files/stopwords_en.txt}:/usr/share/elasticsearch/config/stopwords_en.txt
- ${CL_BASE_DIR:-../../../courtlistener/cl/search/elasticsearch_files/protwords_en.txt}:/usr/share/elasticsearch/config/protwords_en.txt
```

There's no way that `CL_BASE_DIR`, if specified, could satisfy
these usages as well as `/opt/courtlistener`. The intent is clear
that it should stand-in for base directory and not any of its files.
This patch adopts that understanding.