refactor(harvard_merger): rework `process_entry` to be extensible
`import_harvard_pdfs` has a `process_entry` method that is similar,
but different from the entry processing in `process_crosswalk_file`
and abandoned. Since a specialized method for processing an entry
is useful, this patch reconciles the two bits of functionality.

Finally, it adds a `--job` parameter to the command to make adding
other jobs relating to CAP migration easier.