refactor(docker): perform Webpack asset compilation eagerly
While it is simpler to perform asset compilation with Webpack after
the contents of the `cl` directory are copied, doing this comes at
a cost: the layer generated by Webpack is invalidated any time that
the contents of `cl` change. This happens much more frequently than
the compiled `react.js` and its dependencies change.