fix(merge_opinion_versions): support merging of dockets
Until March 2023, every scraped opinion created a new Docket; so, sometimes we will need to merge dockets too. Here we are introducing some logic for merging dockets without losing any of the referencing objects (which is a long list of models) when deleting one of the dockets. Also, update tests displaying the desired behavior