feat(users): Add domain block list
We use a blocklist package on Python to block hundreds of domains from
creating accounts, but they won't update their list as aggresively as
we need. This allows us to keep a list in addition to theirs rather
than try to get them to add things to theirs (they've been closing my
PRs).