fix(alerts): Ensure that percolator pagination returns all alerts
There was a failing test that occasionally misses an alert triggered by the percolator . The issue occurs because, during testing, alerts are created almost simultaneously, which causes some alerts to have the same sorting key for Elasticsearch pagination. In this case, it's better to use the unique ID to sort them instead of date_created.