Set docket alert default order in one place instead of two
Setting a default order at the very beginning of the function for when no
ordering is applied and a separate one halfway into the function for an invalid
ordering is just asking for bugs.  This consolidates the default into one
place, adds a comment to call it out, and ensures that Guido's sacrifice was
not in vain.