Remove the url_replace and sort_caret template tags
By adding a tiny bit of extra context to the template, we can avoid directly
accessing the HttpRequest from the template to replicate data already
calculated in the view.

The url_replace tag was trying to create local links for sorting while
preserving parameters, such as ones that might be used for pagination. However,
the only template to use the tag doesn't use any other parameters.  If this
feature is necessary, then I would argue that the base parameter string should
be calculated once in the view and passed to the template in its context.

The sort_caret tag was likewise accessing the HttpRequest to perform logic best
handled in the view, as was demonstrated by the fact that it needed to be
changed when the view's sorting behavior was changed.  It also embedded HTML
structure that is best left in the template (what if you want to add a CSS
class to highlight the active sort caret?).  The resulting embedded if/else
tags in the template are a bit cumbersome, but if we want to bring it back I
would write it to take one context variable and only what is necessary.