Change alert URLs to use /search-alerts where appropriate
This makes explicit what was implicit, linking directly to the new
profile_search_alerts path when the intention is to load search alerts.  This
is a separate commit because it isn't strictly necessary for the change and
clutters up the diff for review.  Some of these should have no effect
(search/views.py after a user has created a search alert), but others create a
better user experience (alerts/views.py which would otherwise direct users to a
different page when deleting their last alert vs. deleting other alerts).