feat(opinion_page): adds a button to view document attachments
Adds a button to RECAP documents that have or are attachments which allows users
to directly navigate to the parent doc or other attachments without having to
go to the docket.  The button defaults to the next attachment (or original
document if at the last attachment), but a pull down menu allows the selection
of any attachment or the original document.  By grabbing all of the related
documents and filtering them in Python rather than in the database, it incurs
no additional queries and even eliminates some in one edge case.