feat(assets): replace Twitter link with Bluesky in base template
We don't have a Font Awesome Bluesky logo and I spent more time than I wanted
trying to get an SVG to match the other FA icons, so instead I just pulled
forward the social links from the new_base template.  This has the unintended
consequence of replacing the link to free.law with a link to the project's
LinkedIn page, but I think that is okay.
I felt that 24x24 looked the closest to the previous versions.