feat(cl): add example for creating custom settings file
There are file patterns for custom Django settings files in the `.gitignore`,
but there are no examples or documentation for using them and it is not very
discoverable.  I added an example settings file with some documentation and a
couple example settings that solve problems I have run into along with the
corresponding line in the `.env.dev.example` file.

I also updated the wiki with documentation on using custom settings.  Sorry, I
thought it would create a PR or something, I did not know the changes would
immediately go live.  If this gets merged, it should be updated again (happy to
do it).