feat(icons): Add new tiny logo for RECAP
I used the FLP logo as my base, and then optimized it for viewing by
making the word as big as possible. Once that looked good in Inkscape,
I took it to https://optimize.svgomg.net/, which is a phenomenal site
for cleaning up SVGs.

The original will be in the logos repo. This is the optimized version.