refactor(tests): simplify ESIndexTestCase lockfile
The convention for SerializeMixin that Django recommends is to use `__file__`
for the lockfile.  But this doesn't work if multiple classes in a file are
using lockfile.  So I was namespacing the lockfile by appending the class name.
However, this creates a new file in the source directory that that isn't in
the .gitignore.  I could delete the file in tear down, but that isn't 100%
reliable.  The best approach would be to create a random temporary file, but I
need to make sure that parallel runners are forked late enought that they'd use
the same path.

For now, there are no other SerializeMixin test cases using this file as a
lockfile, so we'll use the simplest and recommended approach.