fix(search_templates): prevent RECAP alerts tooltip close button from submitting the form
When a <button> element is inside a <form> and doesn't
have a type, it has an implicit `type="submit"`. Thus,
when hitting Enter in the input, the button caught the
event and prevented the default behavior, so nothing
happened. We now set an explicit `type="button"`, so
the previous behavior is restored: we can now trigger
the search by pressing Enter.