fix(alerts): fixed a test bug
The test test_percolate_document_on_ingestion was relying on a typing bug.  When
creating the RECAPDocument rd_2, it was correctly setting the document_number
to a string, "1".  However, when it changed the value to trigger a webhook, it
set the value to the integer 1.  This worked because 1 != "1", even though
Django converts it later to a string and the result in the database is the
same.