fix(cl): uncrossed the OpinionSearchFunctionalTest streams
You should never cross the streams of TransactionTestCase and TestCase.  All of OpinionSearchFunctionalTest's data objects were being put in a transaction that Selenium couldn't see.  This is fallout from AudioTestCase previously just adding stuff to the live database.  I switched it from TestCase to TransactionTestCase and verified that nothing else importing it was also using TestCase.