refactor(cl): remove SerializeLockFileTestMixin
The reason the parent Django class, SerialMixin, doesn't define the `__lockfile__` is because everything that inherits from it is then using the same lock.  So your Selenium tests would be locked when you're doing ElasticSearch tests, etc.  Better to just use the Django one and deal with `__lockfile__` collisions as they occur.