feat(cl): make --keepdb mean something
This modifies our custom test runner so that tests will never permanently modify
the test database.  This means --keepdb will work even if we use a
TransactionTestCase.

To accomplish this, we continue to let the test runner setup the test database,
running migrations and whatever else it might decide to do in the future.  But
then we create the clones ourselves and erase any trace of the original test
database.

This will always create an extra cone when --parallel is greater than one, but
trying to fix that behavior is messy and I'm fine with an extra 200ms in that
case.