feat(cl): Issue template for reporting bugs with labels
Created an issue template for bug reports.  It requires reporters to select the
component the bug deals with and a severity.  It then applies the appropriate
labels to the issue along with the "bug" and "triage" labels.  It requires the
creation of "severity: trivial", "severity: low", "severity:
medium", "severity: high", and "severity: critical" labels.