fix(ci): Ensure orchestration runs
I noticed in a few of our runs that if we skip earlier steps, the `needs` parameter of the orchestration step makes it get skipped too. That's wrong. 

The idea of the orchestration step is that until it finishes, this action cannot be started a second time, so if it is skipped while a deployment is still happening, that's bad.

The fix here is to make sure that it runs even if earlier deployments are skipped. Adding `if: always()` should accomplish this.