refactor(opinion_page): condensed recap_document and recap_authorities
The view_recap_document and view_recap_authorities were basically the same with
the exception of one expensive query that _authorities has to make.  But
the _authorities version took some shortcuts under the assumption that no one
would directly land on the page and also ignored prayers, so probably best to
unify them.