feat(new_homepage): introduce advanced search in mobile
Current implementation is sub-optimal since we
introduce two different forms for the mobile version
of the new homepage: one shown in the body, and the
other shown in the dialog.

Main concern with this is that any value entered in
the input in the body will not be transferred to the
dialog. This will be addressed in a future commit,
we can probably use two-way binding with Alpine or
hopefully refactor so it's the same form node.