fix(new_homepage): sync entered keyword across forms and screen-sizes
Previously, the mobile version of the search UI in the
homepage had two different forms:

- One in the homepage body.
- One in a dialog.

We now merge both of them, but keep the two different
inputs for keywords. Their values are synced using two-way
binding with Alpine. This way, when a user has entered a
value and then opens the dialog with advanced filters, the
value entered is preserved.

We now also preserve the entered keyword when the user
changes the viewport size. This was previously not
preserved because desktop and mobile also have different
forms. We achieve this using Alpine store.