fix(corpus_search): prevent duplicated fieldset IDs
Since we have more than one instance of the Corpus
Search component rendered at the same time, and we
used static IDs for the scoped fieldsets, these were
clashing and only one of the form elements was being
properly disabled when inactive.

We now use Alpine's 'x-id' directive to generate
different IDs for each instance of the component.
This prevents the leakage of entered data from one
form to the other, and even worse, the override of
filters with the same name.