refactor(cl): removed is_og_bot
I pulled the string while removing is_og_bot from recap_document_context and
discovered that is the only place it is called from, so I removed the whole
thing.

Also removed make_thumb_if_needed, which I forgot to do after inlining it at
the only place from which it was called.