fix(components): close dialogs when crossing mobile breakpoint
Opening an Alpine dialog hides sibling elements from AT by
adding aria-hidden to them and then when the dialog is closed
the attribute is removed. However, if the window is resized
past the breakpoint where the dialog is not visible anymore,
the dialog remains open and thus sibling elements remain
hidden from AT when they shouldn't be.

It can also be really awkward UX for regular users to have a
dialog appear out of nowhere just by resizing the window.

We now close all dialogs when resizing the window past the
mobile breakpoint to avoid this and other potential issues,
as well as improving UX.