fix(html): Re-attach menu to drop down with short descriptions
I noticed that now that the text of the description is smaller, the
dropdown was floated left instead of being linked to the caret.

This re-organizes the HTML to properly put the dropdown at the end of
and within the H3, and then uses pull-right to make the dropdown go
left instead of right.