chore(waffle): change WAFFLE_FLAG_DEFAULT to false
The setting was changed while transitioning to search alerts,
to prevent having to add the flag manually in each test.

However, this is not ideal because we need to always create
flags manually before merging new features, and since the
alerts have been launched and the transition is complete, we
now change it back to Waffle's default behavior.