refactor(dismissible_nav_banner): introduce responsive flag
Previous version of the banner included a keyword `single_row` that
changed the layout of the banner to a row in larger screens, but this
was removed in a recent refactor.

Instead of restoring the exact same keyword, we now change the default
behavior to being always a column (`single_row=False`) and only change
it to a row in larger screens when a new `responsive` flag is passed
with a True value.