fix(dismissible_nav_banner): restore close button behavior
The script to handle closing the banner used a CSS selector that was
dropped in a recent refactor of the banner (class `alert-dismissible`).

However, the class also changed the style of the banner, which broke
layout. We now add a new selector to the script so either that class OR
a new `data-dismissible-alert` attribute can be used. The attribute is
used in the banner, while the class is used in other parts of the code.