fix(search_utils): remove default QuerySet call from signature
The `do_es_search` method was recently refactored with a new
kwarg for courts which has a default in case none is passed.
However, the default is a QuerySet call and it was set in the
method signature, which meant it was defined on module init
instead of on method call, causing tests to fail.
We now use a proper pattern to initialize the default.